CC-380 make changes to tour component
